refactor and create mehod for calculation the deviations for day and site

+++ WeatherParser.Presentation.Entities/WeatherPresentation.cs
@@ -8,8 +8,8 @@ namespace WeatherParser.Presentation.Entities
         //объект для данных о погоде
         public DateTime Date { get; set; }
         public List<double> Temperature { get; set; }
-        public List<int> Pressure { get; set; }
-        public List<int> Humidity { get; set; }
+        public List<double> Pressure { get; set; }
+        public List<double> Humidity { get; set; }
         public List<double> WindSpeed { get; set; }
         public List<string> WindDirection { get; set; }
