+++ WeatherParser.Repository/Entities/WeatherRepository.cs
@@ -12,8 +12,8 @@ namespace WeatherParser.Repository.Entities
         [BsonDateTimeOptions(Kind = DateTimeKind.Local)]
         public DateTime Date { get; set; }
         public List<double> Temperature { get; set; }
-        public List<int> Pressure { get; set; }
-        public List<int> Humidity { get; set; }
+        public List<double> Pressure { get; set; }
+        public List<double> Humidity { get; set; }
         public List<double> WindSpeed { get; set; }
         public List<string> WindDirection { get; set; }
     }
