+++ WeatherParser.Repository/WeatherDataNoSQLRepository.cs
@@ -148,8 +148,8 @@ namespace WeatherParser.Repository
                         var weatherData = new WeatherRepository()
                         {
                             Temperature = new List<double>(),
-                            Humidity = new List<int>(),
-                            Pressure = new List<int>(),
+                            Humidity = new List<double>(),
+                            Pressure = new List<double>(),
                             WindDirection = new List<string>(),
                             WindSpeed = new List<double>()
                         };
