+++ WeatherParser.Service.Entities/WeatherService.cs
@@ -5,8 +5,8 @@
         //объект для сохранения данных о погоде
         public DateTime Date { get; set; }
         public List<double> Temperature { get; set; }
-        public List<int> Pressure { get; set; }
-        public List<int> Humidity { get; set; }
+        public List<double> Pressure { get; set; }
+        public List<double> Humidity { get; set; }
         public List<double> WindSpeed { get; set; }
         public List<string> WindDirection { get; set; }
