+++ WeatherParser.Service.GismeteoService/WeatherDataAngleSharpServiceGismeteo.cs
@@ -47,8 +47,8 @@ namespace WeatherParser.Service.Plugins.GismeteoService
                 WeatherService weatherData = new WeatherService()
                 {
                     Temperature = new List<double>(),
-                    Humidity = new List<int>(),
-                    Pressure = new List<int>(),
+                    Humidity = new List<double>(),
+                    Pressure = new List<double>(),
                     WindDirection = new List<string>(),
                     WindSpeed = new List<double>()
                 };
