+++ WeatherParser.Service.GismeteoService/WeatherDataHtmlAgilityPackServiceGismeteo.cs
@@ -34,8 +34,8 @@ namespace WeatherParser.Service.Plugins.GismeteoService
                 WeatherService listOfWeatherData = new WeatherService()
                 {
                     Temperature = new List<double>(),
-                    Humidity = new List<int>(),
-                    Pressure = new List<int>(),
+                    Humidity = new List<double>(),
+                    Pressure = new List<double>(),
                     WindDirection = new List<string>(),
                     WindSpeed = new List<double>()
                 };
